Revert "Feat/articles"
