+++ package.json
@@ -14,13 +14,11 @@
     "@emotion/styled": "^11.10.4",
     "@fontsource/inter": "^4.5.13",
     "@heroicons/react": "^2.0.11",
-    "dayjs": "^1.11.5",
     "framer-motion": "^7.5.2",
     "next": "12.3.1",
     "next-i18next": "^12.1.0",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "swr": "^1.3.0"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "eslint": "8.24.0",
