+++ pages/index.jsx
@@ -1,36 +1,25 @@
 import Head from "next/head";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { Hero, About, Articles } from "../shared/sections";
-import { useTranslation } from "next-i18next";
+import styles from "../styles/Home.module.css";
+import { Hero } from "../shared/components";
 
 export async function getStaticProps({ locale }) {
   return {
     props: {
-      locale,
-      ...(await serverSideTranslations(locale, [
-        "common",
-        "hero",
-        "about",
-        "home",
-        "articles",
-      ])),
+      ...(await serverSideTranslations(locale, ["common", "hero"])),
     },
   };
 }
-export default function Home({ locale }) {
-  const { t } = useTranslation("home");
-
+export default function Home() {
   return (
     <>
       <Head>
-        <title>Medusa Lab</title>
-        <meta name="description" content={t("meta.description")} />
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
       <Hero />
-      <About />
-      <Articles locale={locale} />
     </>
   );
 }
