--- public/locales/en/about.json
@@ -1,11 +1,0 @@
-{
-  "title": "About us",
-  "camila": {
-    "position": "Developer",
-    "imageAlt": "Profile photo of Camila"
-  },
-  "viviane": {
-    "position": "Senior Developer",
-    "imageAlt": "Profile photo of Viviane"
-  }
-}
