--- public/locales/en/articles.json
@@ -1,5 +1,0 @@
-{
-  "author": "Author",
-  "read": "read",
-  "title": "Articles"
-}
