+++ public/locales/en/common.json
@@ -4,8 +4,7 @@
   "articles": "Articles",
   "projects": "Projects",
   "contact": "Contact",
-  "rights": "medusa.lab. Feito por mulheres.",
+  "rights": "Chakra Templates. All rights reserved",
   "colorMode": "Change color mode",
-  "navigation": "Toggle navigation",
-  "touch": "Get in touch"
+  "navigation": "Toggle navigation"
 }
