--- public/locales/en/home.json
@@ -1,3 +1,0 @@
-{
-  "description": "Welcome to Medusa, a women's tech colective."
-}
