--- public/locales/pt-BR/about.json
@@ -1,11 +1,0 @@
-{
-  "title": "Sobre nós",
-  "camila": {
-    "position": "Desenvolvedora",
-    "imageAlt": "Foto de perfil da Camila"
-  },
-  "viviane": {
-    "position": "Desenvolvedora Sênior",
-    "imageAlt": "Foto de perfil da Viviane"
-  }
-}
