--- public/locales/pt-BR/articles.json
@@ -1,5 +1,0 @@
-{
-  "author": "Autora",
-  "read": "de leitura",
-  "title": "Artigos"
-}
