+++ public/locales/pt-BR/common.json
@@ -4,8 +4,7 @@
   "articles": "Artigos",
   "projects": "Projetos",
   "contact": "Contato",
-  "rights": "medusa.lab. Feito por mulheres.",
+  "rights": "Chakra Templates. Todos os direitos reservados",
   "colorMode": "Mudar cor do tema",
-  "navigation": "Mudar navegação",
-  "touch": "Entre em contato"
+  "navigation": "Mudar navegação"
 }
