+++ public/locales/pt-BR/hero.json
@@ -1,5 +1,5 @@
 {
   "title": "Olá, Viviane & Camila aqui",
-  "subtitle": "Somos desenvolvedoras apaixonadas por construir ferramentas que empoderam pessoas.",
+  "subtitle": "Somos desenvolvedoras apaixonadas por construir ferramentas que empoderem pessoas",
   "about": "Sobre nós"
 }
