--- public/locales/pt-BR/home.json
@@ -1,3 +1,0 @@
-{
-  "description": "Bem-vinda a Medusa, um coletivo de mulheres tecnologistas."
-}
