+++ shared/components/Footer.jsx
@@ -33,7 +33,6 @@ export default function SmallWithNavigation() {
       borderTop={1}
       borderStyle={"solid"}
       borderColor={useColorModeValue("gray.200", "gray.900")}
-      mt={10}
     >
       <Container
         as={Stack}
