+++ shared/components/index.js
@@ -1,2 +1,3 @@
-export { default as SectionLayout } from "./SectionLayout";
-export { default as Image } from "./Image";
+export { default as Header } from "./Header";
+export { default as Footer } from "./Footer";
+export { default as Hero } from "./Hero";
