--- shared/sections/index.js
@@ -1,5 +1,0 @@
-export { default as Header } from "../sections/Header";
-export { default as Footer } from "../sections/Footer";
-export { default as Hero } from "./Hero";
-export { default as About } from "./About";
-export { default as Articles } from "./Articles";
